# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Contact {
  id: number;
  name: string;
  email?: string | null;
  phone_number?: string | null;
  avatar_url?: string | null;
}

export type ConversationMeta = {
  sender: Contact;
  assignee?: Contact | null;
};

export interface Message {
  id: number;
  content?: string | null;
  message_type: number;
  created_at: number;
  sender?: Contact | null;
}

export interface Conversation {
  id: number;
  status: string;
  unread_count: number;
  last_activity_at?: number | null;
  meta: ConversationMeta;
  messages?: Message[];
}

export interface SendMessageRequest {
  conversationId: number;
  content: string;
}

export interface ToggleStatusRequest {
  conversationId: number;
  status: string;
}