# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  name: string;
  email?: string | null;
  phone_number?: string | null;
  avatar_url?: string | null;
}