# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from "./contact";

export interface Message {
  id: number;
  content?: string | null;
  message_type: number;
  created_at: number;
  sender?: Contact | null;
}