# lib/api-zod/src/generated/types/toggleStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ToggleStatusRequest {
  conversationId: number;
  status: string;
}